new data from 45degs only and 50 trajectories
The purpose of this data is to see if the model can learn a single trajectory if there is enough data on that trajectory. The reason for this experiment was because it seemed that the task space was much too large and required too much data, which meant that nearly every trajectory was an outlier relative to the others. The task space needed narrowing so that lots of data could be collected from a similar region of the task space.